# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Teemer Moving & Storage API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /customer/quotes/{id}:
    get:
      operationId: getCustomerQuote
      tags: [customer]
      summary: Get one of the customer's saved quotes
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Quote
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteResponse"
        "404":
          description: Not found
    patch:
      operationId: updateCustomerQuote
      tags: [customer]
      summary: Update a saved quote (only while still in quote_requested status)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerQuoteUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteResponse"
        "409":
          description: No longer editable
...
  /customer/jobs/{id}:
    get:
      operationId: getCustomerJob
      tags: [customer]
      summary: Get full detail for a customer job
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Job detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerJobDetail"
  /customer/jobs/{id}/contract.pdf:
    get:
      operationId: downloadCustomerContractPdf
      tags: [customer]
      summary: Download contract PDF for the customer's job
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: PDF
          content:
            application/pdf:
              schema: { type: string, format: binary }
  /customer/invoices/{jobId}/pdf:
    get:
      operationId: downloadCustomerInvoicePdf
      tags: [customer]
      summary: Download invoice PDF
      parameters:
        - in: path
          name: jobId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: PDF
          content:
            application/pdf:
              schema: { type: string, format: binary }
...
  /customer/payment-requests/{id}:
    get:
      operationId: getCustomerPaymentRequest
      tags: [customer]
      summary: Get one of the customer's payment requests
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Payment request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerPaymentRequest"
  /customer/payment-requests/{id}/checkout:
    post:
      operationId: payCustomerPaymentRequest
      tags: [customer]
      summary: Start Stripe checkout for a payment request
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Checkout session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutSessionResponse"
  /customer/jobs/{id}/balance-checkout:
    post:
      operationId: payCustomerJobBalance
      tags: [customer]
      summary: Start Stripe checkout for the customer's outstanding balance
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Checkout session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutSessionResponse"

  # ─── Admin: payment requests + payments view (Task #44) ───────────────────
...
components:
  schemas:
...
    InventoryMap:
      type: object
      description: Map of item name to quantity. Keys are item names, values are quantities.
      additionalProperties:
        type: integer
      example:
        "Sofa": 1
        "Queen bed": 2
        "Refrigerator": 1

    # ─── Quote Request ──────────────────────────────────────────────────────────
...
    QuoteResponse:
      type: object
      properties:
        id:
          type: string
        status:
          type: string
          enum: [quote_requested, deposit_paid, booked]
          description: "quote_requested | deposit_paid | booked"
        createdAt:
          type: string
        # Pricing results
        crewSize:
          type: integer
        hourlyRate:
          type: number
        estimatedHours:
          type: number
        laborSubtotal:
          type: number
        materialsSubtotal:
          type: number
        pianoSurcharge:
          type: number
        commercialAdjustment:
          type: number
          description: "Amount added above residential estimate for commercial moves"
        mountedTVFee:
          type: number
          description: "Per-TV wall-mounted TV dismount/remount surcharge (count × $50/TV); included in totalEstimate"
        totalEstimate:
          type: number
        depositAmount:
          type: number
        serviceType:
          type: string
          enum: [moving, junk_removal]
          description: "Service type: moving or junk_removal"
        junkLoadSize:
          type: string
          enum: [small, medium, large, full_truck]
        junkStairsFlights:
          type: integer
        junkHeavyItemsCount:
          type: integer
        junkConstructionDebris:
          type: boolean
        junkSameDay:
          type: boolean
        junkHazardousItems:
          type: boolean
        junkBasePrice:
          type: number
        junkAddonsTotal:
          type: number
        # Legacy compat
        estimatedPriceLow:
          type: number
        estimatedPriceHigh:
          type: number
        # Booking detail confirmations (echoed back so the deposit page can
        # surface them for the customer)
        packingDate:
          type: string
          nullable: true
        packingArrivalWindow:
          type: string
          nullable: true
        hasMountedTVs:
          type: boolean
        mountedTVCount:
          type: integer
        parkingInstructions:
          type: string
          nullable: true
        # Full request echo
        quoteRequest:
          $ref: "#/components/schemas/QuoteRequestEcho"
      required:
        - id
        - status
        - createdAt
        - quoteRequest
    QuoteRequestEcho:
      type: object
      description: Echo of the quote request fields as stored
      properties:
        contactName:
          type: string
        phone:
          type: string
        email:
          type: string
        moveDate:
          type: string
        arrivalTimeWindow:
          type: string
        pickupAddress:
          type: string
        dropoffAddress:
          type: string
        secondStop:
          type: string
        storageNeeded:
          type: boolean
        storageUnitChoice:
          type: string
        additionalNotes:
          type: string
        moveType:
          type: string
        residentialOrCommercial:
          type: string
        originAddress:
          type: string
        destinationAddress:
          type: string
        numberOfBedrooms:
          type: integer
        numberOfLivingRooms:
          type: integer
        isFullyFurnished:
          type: boolean
        hasGarage:
          type: boolean
        hasOutdoorFurniture:
          type: boolean
        hasStairs:
          type: boolean
        hasHeavyItems:
          type: boolean
        pianoType:
          type: string
          enum: [none, upright, grand]
        pianoFloor:
          type: string
          enum: [ground, stairs]
        isCommercial:
          type: boolean
        commercialBusinessType:
          type: string
        commercialSizeTier:
          type: string
          enum: [small, medium, large, enterprise]
        inventory:
          $ref: "#/components/schemas/InventoryMap"
        boxesAlreadyPacked:
          type: integer
        needsPackingMaterials:
          type: boolean
        smallBoxes:
          type: integer
        mediumBoxes:
          type: integer
        serviceType:
          type: string
          enum: [moving, junk_removal]
        junkLoadSize:
          type: string
          enum: [small, medium, large, full_truck]
        junkStairsFlights:
          type: integer
        junkHeavyItemsCount:
          type: integer
        junkConstructionDebris:
          type: boolean
        junkSameDay:
          type: boolean
        junkHazardousItems:
          type: boolean
        moveSize:
          type: string
        numberOfRooms:
          type: integer
        packingHelpNeeded:
          type: string
        specialItems:
          type: string
...
    CustomerQuoteUpdate:
      type: object
      properties:
        moveDate: { type: string }
        arrivalTimeWindow: { type: string }
        pickupAddress: { type: string }
        dropoffAddress: { type: string }
        secondStop: { type: string }
        additionalNotes: { type: string }
        parkingInstructions: { type: string }
        phone: { type: string }
...
    CustomerJobDetail:
      type: object
      properties:
        job: { type: object }
        quote: { type: object, nullable: true }
        invoice: { type: object, nullable: true }
        payments:
          type: array
          items: { type: object }
...
    CustomerPaymentRequest:
      type: object
      properties:
        id: { type: integer }
        customerId: { type: integer }
        amountCents: { type: integer }
        description: { type: string }
        status: { type: string }
        stripeSessionId: { type: string, nullable: true }
        confirmationNumber: { type: string, nullable: true }
        createdAt: { type: string, nullable: true }
        paidAt: { type: string, nullable: true }
        payUrl: { type: string, nullable: true }
    CheckoutSessionResponse:
      type: object
      properties:
        url: { type: string, nullable: true }
        sessionId: { type: string }
...