# lib/api-zod/src/generated/types/customerJobDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Teemer Moving & Storage API
 * OpenAPI spec version: 0.1.0
 */
import type { CustomerJobDetailInvoice } from "./customerJobDetailInvoice";
import type { CustomerJobDetailJob } from "./customerJobDetailJob";
import type { CustomerJobDetailPaymentsItem } from "./customerJobDetailPaymentsItem";
import type { CustomerJobDetailQuote } from "./customerJobDetailQuote";

export interface CustomerJobDetail {
  job?: CustomerJobDetailJob;
  quote?: CustomerJobDetailQuote;
  invoice?: CustomerJobDetailInvoice;
  payments?: CustomerJobDetailPaymentsItem[];
}